refactor: consolidate codebase and remove unused code
Major simplification effort reducing ~2,800 lines of code:

**Task Store (10+ files → 3 files)**
- Consolidated scattered actions into taskStore.filters.ts and taskStore.import.ts
- Removed redundant re-export wrappers and type files
- Simplified state management while preserving all functionality

**Validation Module (7 files → 1 file)**
- Merged schemas, validators, and sanitization into single validation.ts
- Eliminated unnecessary directory structure

**Conflict Resolution (6 files → 1 file)**
- Consolidated board, task, and settings conflict handlers
- Simplified import conflict resolution logic

**Accessibility Module (6 files → 1 file)**
- Combined focus management, keyboard navigation, announcements, and settings
- Preserved WCAG 2.1 AA compliance functionality

**Error Handling Infrastructure (deleted)**
- Removed 413 lines of unused error handling code (AppError, ErrorRecovery, etc.)
- Infrastructure was never integrated into the application

**Dialog State Boilerplate**
- Replaced 6 separate useState calls with single activeDialog state
- Reduced Sidebar props from 12 to 2

**Security Module**
- Removed unused CSP generation code (CloudFront handles CSP headers)
