+++ src/lib/stores/taskStore.import.ts
@@ -1,5 +1,5 @@
 /**
- * Import/export operations for task store
+ * Task store import/export operations
  * Handles bulk data operations and integration with import/export utilities
  */
 
@@ -7,24 +7,22 @@ import { Task } from '@/lib/types';
 import { taskDB } from '@/lib/utils/database';
 import { exportTasks as exportTasksUtil, ExportData } from '@/lib/utils/exportImport';
 
-// Type for store state access
-type ImportExportStoreState = {
+// Type for store state access - minimal interface needed by import/export
+type ImportExportState = {
   tasks: Task[];
-  filters: { boardId?: string; crossBoardSearch?: boolean };
   applyFilters: () => Promise<void>;
 };
 
-// Type for store setter - using any to avoid complex type inference issues
-type StoreSetter = (state: any) => void;
+// Using any for setter to avoid complex type inference issues with Zustand
+// This is a pragmatic choice - the actual type safety comes from the store definition
+type StoreSetter = (partial: any) => void;
 
 /**
  * Exports tasks to ExportData format
- * Supports filtering by board and including/excluding archived tasks
  */
-export function createExportTasks(get: () => ImportExportStoreState) {
+export function createExportTasks(get: () => ImportExportState) {
   return (options = { includeArchived: true }): ExportData => {
-    const { tasks } = get();
-    return exportTasksUtil(tasks, options);
+    return exportTasksUtil(get().tasks, options);
   };
 }
 
@@ -32,44 +30,27 @@ export function createExportTasks(get: () => ImportExportStoreState) {
  * Imports tasks from external source
  * Updates existing tasks or adds new ones based on ID
  */
-export function createImportTasks(
-  get: () => ImportExportStoreState,
-  set: StoreSetter
-) {
+export function createImportTasks(get: () => ImportExportState, set: StoreSetter) {
   return async (tasks: Task[]) => {
     try {
       set({ isLoading: true, error: null });
 
-      const { tasks: existingTasks } = get();
-      const existingIds = new Set(existingTasks.map(t => t.id));
+      const existingIds = new Set(get().tasks.map(t => t.id));
 
-      // Add or update tasks in database
       for (const task of tasks) {
         if (existingIds.has(task.id)) {
-          // Update existing task
           await taskDB.updateTask(task);
         } else {
-          // Add new task
           await taskDB.addTask(task);
         }
       }
 
-      // Update store state
-      set((state: any) => {
-        const taskMap = new Map(state.tasks.map((t: Task) => [t.id, t]));
-
-        // Add or update imported tasks
-        tasks.forEach(task => {
-          taskMap.set(task.id, task);
-        });
-
-        return {
-          tasks: Array.from(taskMap.values()),
-          isLoading: false,
-        };
+      set((state: { tasks: Task[] }) => {
+        const taskMap = new Map(state.tasks.map(t => [t.id, t]));
+        tasks.forEach(task => taskMap.set(task.id, task));
+        return { tasks: Array.from(taskMap.values()), isLoading: false };
       });
 
-      // Reapply filters after import
       void get().applyFilters();
     } catch (error: unknown) {
       set({
@@ -85,32 +66,22 @@ export function createImportTasks(
  * Bulk adds tasks to the store and database
  * Processes in batches to avoid overwhelming the database
  */
-export function createBulkAddTasks(
-  get: () => ImportExportStoreState,
-  set: StoreSetter
-) {
+export function createBulkAddTasks(get: () => ImportExportState, set: StoreSetter) {
   return async (tasks: Task[]) => {
     try {
       set({ isLoading: true, error: null });
 
-      // Process tasks in batches to avoid overwhelming the database
       const batchSize = 50;
-      const batches = [];
       for (let i = 0; i < tasks.length; i += batchSize) {
-        batches.push(tasks.slice(i, i + batchSize));
-      }
-
-      for (const batch of batches) {
+        const batch = tasks.slice(i, i + batchSize);
         await Promise.all(batch.map(task => taskDB.addTask(task)));
       }
 
-      // Update store state
-      set((state: any) => ({
+      set((state: { tasks: Task[] }) => ({
         tasks: [...state.tasks, ...tasks],
-        isLoading: false,
+        isLoading: false
       }));
 
-      // Reapply filters after bulk add
       void get().applyFilters();
     } catch (error: unknown) {
       set({
