+++ src/lib/stores/taskStore.ts
@@ -1,29 +1,23 @@
 /**
  * Main task store - Consolidated architecture
- * Simplified from 7 files to 4 files for better maintainability
- *
- * File structure:
- * - taskStore.ts (this file) - Main store composition
- * - taskStore.actions.ts - CRUD, filters, and search operations
- * - taskStore.validation.ts - Validation and error recovery
- * - taskStore.importExport.ts - Import/export operations (unchanged)
+ * Simplified to 3 files for better maintainability:
+ * - taskStore.ts (this file) - Main store with CRUD and validation
+ * - taskStore.filters.ts - Filtering, search, and related operations
+ * - taskStore.import.ts - Import/export operations
  */
 
 import { create } from 'zustand';
 import { devtools } from 'zustand/middleware';
 import { Task, TaskFilters, SearchState, SearchScope } from '@/lib/types';
+import { taskDB } from '@/lib/utils/database';
+import { sanitizeTaskData } from '@/lib/utils/security';
+import { validateTaskIntegrity } from '@/lib/utils/taskValidation';
 import { ExportData } from '@/lib/utils/exportImport';
 import { type SearchCache } from '@/lib/utils/taskFiltering';
 
-// Import action creators from consolidated modules
+// Import from consolidated filter module
 import {
-  createAddTask,
-  createUpdateTask,
-  createDeleteTask,
-  createMoveTask,
-  createMoveTaskToBoard,
-  createArchiveTask,
-  createUnarchiveTask,
+  applyFiltersToTasks,
   createSetFilters,
   createSetBoardFilter,
   createSetCrossBoardSearch,
@@ -34,24 +28,20 @@ import {
   createSetHighlightedTask,
   createNavigateToTaskBoard,
   createLoadSearchPreferences,
-  createSaveSearchScope
-} from './taskStore.actions';
+  createSaveSearchScope,
+} from './taskStore.filters';
 
-import {
-  createValidateBoardAccess,
-  createHandleBoardDeletion,
-  createRecoverFromSearchError,
-  createInitializeStore,
-  validateTaskIntegrity
-} from './taskStore.validation';
-
+// Import from consolidated import module
 import {
   createExportTasks,
   createImportTasks,
   createBulkAddTasks
-} from './taskStoreImportExport';
+} from './taskStore.import';
 
-// State interface
+// ============================================================================
+// State & Action Interfaces
+// ============================================================================
+
 interface TaskState {
   tasks: Task[];
   filteredTasks: Task[];
@@ -63,7 +53,6 @@ interface TaskState {
   searchCache: SearchCache;
 }
 
-// Actions interface
 interface TaskActions {
   // State management
   setTasks: (tasks: Task[]) => void;
@@ -89,8 +78,6 @@ interface TaskActions {
   // Filtering and search
   applyFilters: () => Promise<void>;
   clearFilters: () => void;
-
-  // Search state management
   setHighlightedTask: (taskId: string | undefined) => void;
   clearSearch: () => void;
   navigateToTaskBoard: (taskId: string) => Promise<{ success: boolean; boardId?: string; error?: string }>;
@@ -99,23 +86,26 @@ interface TaskActions {
   loadSearchPreferences: () => Promise<void>;
   saveSearchScope: (scope: SearchScope) => Promise<void>;
 
-  // Error handling and validation
+  // Validation
   validateBoardAccess: (boardId: string) => Promise<boolean>;
   handleBoardDeletion: (deletedBoardId: string) => void;
   recoverFromSearchError: () => void;
   validateTaskIntegrity: (task: Task) => boolean;
 
-  // Import/Export operations
+  // Import/Export
   exportTasks: (options?: { includeArchived: boolean; boardIds?: string[] }) => ExportData;
   importTasks: (tasks: Task[]) => Promise<void>;
   bulkAddTasks: (tasks: Task[]) => Promise<void>;
 
-  // Store initialization
+  // Initialization
   initializeStore: () => Promise<void>;
 }
 
-// Use satisfies operator (TS 5.0+) for type-safe initial state
-const initialState = {
+// ============================================================================
+// Initial State
+// ============================================================================
+
+const initialState: TaskState = {
   tasks: [],
   filteredTasks: [],
   filters: {
@@ -124,19 +114,276 @@ const initialState = {
     crossBoardSearch: false,
   },
   searchState: {
-    scope: 'current-board' as const,
+    scope: 'current-board',
     highlightedTaskId: undefined,
   },
   isLoading: false,
   isSearching: false,
   error: null,
   searchCache: new Map(),
-} satisfies TaskState;
+};
 
-/**
- * Main task store
- * Composed from focused modules for better maintainability
- */
+// ============================================================================
+// CRUD Operations (inline for simplicity)
+// ============================================================================
+
+function createAddTask(get: () => TaskState & TaskActions, set: (state: Partial<TaskState>) => void) {
+  return async (taskData: Omit<Task, 'id' | 'createdAt' | 'updatedAt'>) => {
+    try {
+      set({ isLoading: true, error: null });
+
+      const sanitizedData = sanitizeTaskData({
+        title: taskData.title,
+        description: taskData.description,
+        tags: taskData.tags,
+      });
+
+      if (!sanitizedData.title.trim()) {
+        throw new Error('Task title is required');
+      }
+
+      const newTask: Task = {
+        ...taskData,
+        ...sanitizedData,
+        id: crypto.randomUUID(),
+        createdAt: new Date(),
+        updatedAt: new Date(),
+      };
+
+      await taskDB.addTask(newTask);
+
+      const { tasks, filters } = get();
+      const updatedTasks = [...tasks, newTask];
+
+      set({
+        tasks: updatedTasks,
+        filteredTasks: applyFiltersToTasks(updatedTasks, filters),
+        isLoading: false,
+        searchCache: new Map(),
+      });
+    } catch (error: unknown) {
+      set({
+        error: error instanceof Error ? error.message : 'Failed to add task',
+        isLoading: false
+      });
+    }
+  };
+}
+
+function createUpdateTask(get: () => TaskState & TaskActions, set: (state: Partial<TaskState>) => void) {
+  return async (taskId: string, updates: Partial<Task>) => {
+    try {
+      set({ isLoading: true, error: null });
+
+      const task = get().tasks.find(t => t.id === taskId);
+      if (!task) throw new Error('Task not found');
+
+      const updatedTask = { ...task, ...updates, updatedAt: new Date() };
+      await taskDB.updateTask(updatedTask);
+
+      const { tasks, filters } = get();
+      const updatedTasks = tasks.map(t => t.id === taskId ? updatedTask : t);
+
+      set({
+        tasks: updatedTasks,
+        filteredTasks: applyFiltersToTasks(updatedTasks, filters),
+        isLoading: false,
+        searchCache: new Map(),
+      });
+    } catch (error: unknown) {
+      set({
+        error: error instanceof Error ? error.message : 'Failed to update task',
+        isLoading: false
+      });
+    }
+  };
+}
+
+function createDeleteTask(get: () => TaskState & TaskActions, set: (state: Partial<TaskState>) => void) {
+  return async (taskId: string) => {
+    try {
+      set({ isLoading: true, error: null });
+      await taskDB.deleteTask(taskId);
+
+      const { tasks, filters } = get();
+      const updatedTasks = tasks.filter(t => t.id !== taskId);
+
+      set({
+        tasks: updatedTasks,
+        filteredTasks: applyFiltersToTasks(updatedTasks, filters),
+        isLoading: false,
+        searchCache: new Map(),
+      });
+    } catch (error: unknown) {
+      set({
+        error: error instanceof Error ? error.message : 'Failed to delete task',
+        isLoading: false
+      });
+    }
+  };
+}
+
+function createMoveTask(get: () => TaskState & TaskActions, set: (state: Partial<TaskState>) => void) {
+  return async (taskId: string, newStatus: Task['status']) => {
+    const task = get().tasks.find(t => t.id === taskId);
+    if (!task) return;
+
+    const updates: Partial<Task> = { status: newStatus, updatedAt: new Date() };
+
+    if (newStatus === 'done') {
+      updates.progress = 100;
+      if (!task.completedAt) updates.completedAt = new Date();
+    } else if (newStatus === 'in-progress') {
+      if (task.status === 'todo') updates.progress = 0;
+      updates.completedAt = undefined;
+    } else if (newStatus === 'todo') {
+      updates.progress = undefined;
+      updates.completedAt = undefined;
+    }
+
+    try {
+      await get().updateTask(taskId, updates);
+    } catch (error: unknown) {
+      set({ error: error instanceof Error ? error.message : 'Failed to move task' });
+    }
+  };
+}
+
+function createMoveTaskToBoard(get: () => TaskState & TaskActions, set: (state: Partial<TaskState>) => void) {
+  return async (taskId: string, targetBoardId: string) => {
+    try {
+      set({ isLoading: true, error: null });
+
+      const task = get().tasks.find(t => t.id === taskId);
+      if (!task) throw new Error('Task not found');
+      if (task.boardId === targetBoardId) throw new Error('Task is already on this board');
+
+      await get().updateTask(taskId, { boardId: targetBoardId, updatedAt: new Date() });
+      await get().applyFilters();
+    } catch (error: unknown) {
+      set({ error: error instanceof Error ? error.message : 'Failed to move task to board' });
+      throw error;
+    } finally {
+      set({ isLoading: false });
+    }
+  };
+}
+
+function createArchiveTask(get: () => TaskState & TaskActions, set: (state: Partial<TaskState>) => void) {
+  return async (taskId: string) => {
+    try {
+      await get().updateTask(taskId, { archivedAt: new Date(), updatedAt: new Date() });
+    } catch (error: unknown) {
+      set({ error: error instanceof Error ? error.message : 'Failed to archive task' });
+    }
+  };
+}
+
+function createUnarchiveTask(get: () => TaskState & TaskActions, set: (state: Partial<TaskState>) => void) {
+  return async (taskId: string) => {
+    try {
+      await get().updateTask(taskId, { archivedAt: undefined, updatedAt: new Date() });
+    } catch (error: unknown) {
+      set({ error: error instanceof Error ? error.message : 'Failed to unarchive task' });
+    }
+  };
+}
+
+// ============================================================================
+// Validation Operations (inline for simplicity)
+// ============================================================================
+
+function createValidateBoardAccess() {
+  return async (boardId: string): Promise<boolean> => {
+    try {
+      const boards = await taskDB.getBoards();
+      const board = boards.find(b => b.id === boardId);
+      return board !== undefined && !board.archivedAt;
+    } catch {
+      return false;
+    }
+  };
+}
+
+function createHandleBoardDeletion(get: () => TaskState, set: (state: Partial<TaskState>) => void) {
+  return (deletedBoardId: string) => {
+    const { tasks, filters, filteredTasks } = get();
+
+    try {
+      const updatedTasks = tasks.filter(task => task.boardId !== deletedBoardId);
+      const updatedFilteredTasks = filteredTasks.filter(task => task.boardId !== deletedBoardId);
+      const updatedFilters = filters.boardId === deletedBoardId
+        ? { ...filters, boardId: undefined }
+        : filters;
+
+      set({
+        tasks: updatedTasks,
+        filteredTasks: updatedFilteredTasks,
+        filters: updatedFilters,
+        error: null
+      });
+    } catch {
+      set({ error: 'Failed to update tasks after board deletion. Please refresh the page.' });
+    }
+  };
+}
+
+function createRecoverFromSearchError(get: () => TaskState & TaskActions, set: (state: Partial<TaskState>) => void) {
+  return () => {
+    try {
+      set({
+        filters: { ...get().filters, search: '' },
+        isSearching: false,
+        error: null,
+      });
+      void get().applyFilters();
+    } catch {
+      set({
+        filteredTasks: get().tasks,
+        filters: { search: '', tags: [], crossBoardSearch: false },
+        isSearching: false,
+        error: 'Search functionality temporarily unavailable. Please refresh the page.',
+      });
+    }
+  };
+}
+
+function createInitializeStore(get: () => TaskState & TaskActions, set: (state: Partial<TaskState>) => void) {
+  return async () => {
+    try {
+      set({ isLoading: true, error: null });
+
+      if (typeof window === 'undefined') {
+        set({ isLoading: false });
+        return;
+      }
+
+      await taskDB.init();
+      const tasks = await taskDB.getTasks();
+
+      const processedTasks = tasks.map(task => ({
+        ...task,
+        createdAt: new Date(task.createdAt),
+        updatedAt: new Date(task.updatedAt),
+        completedAt: task.completedAt ? new Date(task.completedAt) : undefined,
+        archivedAt: task.archivedAt ? new Date(task.archivedAt) : undefined,
+      }));
+
+      set({ tasks: processedTasks, filteredTasks: processedTasks, isLoading: false });
+      await get().loadSearchPreferences();
+    } catch (error: unknown) {
+      set({
+        error: error instanceof Error ? error.message : 'Failed to initialize store',
+        isLoading: false
+      });
+    }
+  };
+}
+
+// ============================================================================
+// Store Definition
+// ============================================================================
+
 export const useTaskStore = create<TaskState & TaskActions>()(
   devtools(
     (set, get) => ({
@@ -150,7 +397,7 @@ export const useTaskStore = create<TaskState & TaskActions>()(
       setError: (error) => set({ error }),
       clearSearchCache: () => set({ searchCache: new Map() }),
 
-      // CRUD operations (from taskStoreActions)
+      // CRUD operations
       addTask: createAddTask(get, set),
       updateTask: createUpdateTask(get, set),
       deleteTask: createDeleteTask(get, set),
@@ -159,7 +406,7 @@ export const useTaskStore = create<TaskState & TaskActions>()(
       archiveTask: createArchiveTask(get, set),
       unarchiveTask: createUnarchiveTask(get, set),
 
-      // Filter operations (from taskStoreFilters)
+      // Filter operations (from taskStore.filters.ts)
       setFilters: createSetFilters(get, set),
       setBoardFilter: createSetBoardFilter(get, set),
       setCrossBoardSearch: createSetCrossBoardSearch(get, set),
@@ -168,18 +415,18 @@ export const useTaskStore = create<TaskState & TaskActions>()(
       clearFilters: createClearFilters(get, set),
       clearSearch: createClearSearch(get, set),
 
-      // Search operations (from taskStoreSearch)
+      // Search operations (from taskStore.filters.ts)
       setHighlightedTask: createSetHighlightedTask(set),
       navigateToTaskBoard: createNavigateToTaskBoard(get, set),
       loadSearchPreferences: createLoadSearchPreferences(get, set),
       saveSearchScope: createSaveSearchScope(),
 
-      // Import/Export operations (from taskStoreImportExport)
+      // Import/Export operations (from taskStore.import.ts)
       exportTasks: createExportTasks(get),
       importTasks: createImportTasks(get, set),
       bulkAddTasks: createBulkAddTasks(get, set),
 
-      // Validation and error handling (from taskStoreValidation)
+      // Validation operations
       validateBoardAccess: createValidateBoardAccess(),
       handleBoardDeletion: createHandleBoardDeletion(get, set),
       recoverFromSearchError: createRecoverFromSearchError(get, set),
