--- src/lib/utils/validation/index.ts
@@ -1,37 +1,0 @@
-/**
- * Validation module - re-exports all validation utilities
- */
-
-// Types
-export type {
-  ValidationSchema,
-  DetailedValidationResult,
-  ValidationError,
-  ValidationWarning,
-  SanitizationOptions
-} from './types';
-
-// Schemas
-export {
-  taskSchema,
-  boardSchema,
-  settingsSchema,
-  exportDataSchema
-} from './schemas';
-
-// Schema validation
-export { validateSchema } from './schemaValidator';
-
-// Sanitization
-export { sanitizeData, getDefaultValue } from './sanitization';
-
-// Entity validators
-export {
-  validateTask,
-  validateBoard,
-  validateSettings,
-  validateExportData
-} from './entityValidators';
-
-// Relationship validation
-export { validateDataRelationships } from './relationshipValidator';
